refactor: merge revisions into one single file
