--- src/api/database/alembic/versions/0c9f61bb86b0_increase_telephone_length.py
@@ -1,42 +1,0 @@
-"""increase telephone length
-
-Revision ID: 0c9f61bb86b0
-Revises: 19b4f3cd7497
-Create Date: 2024-06-16 13:45:08.293069
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = "0c9f61bb86b0"
-down_revision: Union[str, None] = "19b4f3cd7497"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "alunos",
-        "telefone",
-        existing_type=mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=20),
-        type_=sa.String(length=23),
-        existing_nullable=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "alunos",
-        "telefone",
-        existing_type=sa.String(length=23),
-        type_=mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=20),
-        existing_nullable=False,
-    )
-    # ### end Alembic commands ###
