--- src/api/database/alembic/versions/1660a68fc219_fix_table_relationships.py
@@ -1,32 +1,0 @@
-"""fix table relationships
-
-Revision ID: 1660a68fc219
-Revises: 40e5b9c282a1
-Create Date: 2024-06-17 02:57:07.550497
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "1660a68fc219"
-down_revision: Union[str, None] = "40e5b9c282a1"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint("usuarios_ibfk_1", "usuarios", type_="foreignkey")
-    op.create_foreign_key(None, "usuarios", "tipo_usuario", ["tipo_usuario_id"], ["id"])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, "usuarios", type_="foreignkey")
-    op.create_foreign_key(
-        "usuarios_ibfk_1", "usuarios", "usuarios", ["tipo_usuario_id"], ["id"]
-    )
-    # ### end Alembic commands ###
