--- src/api/database/alembic/versions/40e5b9c282a1_fix_table_relationships.py
@@ -1,46 +1,0 @@
-"""fix table relationships
-
-Revision ID: 40e5b9c282a1
-Revises: c15640acc34d
-Create Date: 2024-06-17 02:51:31.240539
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "40e5b9c282a1"
-down_revision: Union[str, None] = "c15640acc34d"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("alunos", sa.Column("usuario_id", sa.Integer(), nullable=False))
-    op.create_foreign_key(None, "alunos", "usuarios", ["usuario_id"], ["id"])
-    op.create_index(
-        op.f("ix_professores_usuario_id"), "professores", ["usuario_id"], unique=False
-    )
-    op.create_index(
-        op.f("ix_solicitacoes_professor_id"),
-        "solicitacoes",
-        ["professor_id"],
-        unique=False,
-    )
-    op.create_index(op.f("ix_tarefas_aluno_id"), "tarefas", ["aluno_id"], unique=False)
-    op.create_foreign_key(None, "usuarios", "usuarios", ["tipo_usuario_id"], ["id"])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, "usuarios", type_="foreignkey")
-    op.drop_index(op.f("ix_tarefas_aluno_id"), table_name="tarefas")
-    op.drop_index(op.f("ix_solicitacoes_professor_id"), table_name="solicitacoes")
-    op.drop_index(op.f("ix_professores_usuario_id"), table_name="professores")
-    op.drop_constraint(None, "alunos", type_="foreignkey")
-    op.drop_column("alunos", "usuario_id")
-    # ### end Alembic commands ###
