--- src/api/database/alembic/versions/713e6a02abcb_adicionado_descricao_para_tipo_usuario.py
@@ -1,35 +1,0 @@
-"""adicionado descricao para tipo usuario
-
-Revision ID: 713e6a02abcb
-Revises: 1660a68fc219
-Create Date: 2024-06-17 03:21:08.863132
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "713e6a02abcb"
-down_revision: Union[str, None] = "1660a68fc219"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "tipo_usuario", sa.Column("descricao", sa.String(length=255), nullable=False)
-    )
-    op.create_index(
-        op.f("ix_tipo_usuario_descricao"), "tipo_usuario", ["descricao"], unique=False
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_tipo_usuario_descricao"), table_name="tipo_usuario")
-    op.drop_column("tipo_usuario", "descricao")
-    # ### end Alembic commands ###
