+++ src/api/database/alembic/versions/7b60b08c6475_first_commit.py
@@ -1,0 +1,252 @@
+"""first commit
+
+Revision ID: 7b60b08c6475
+Revises:
+Create Date: 2024-06-17 09:39:16.336523
+
+"""
+from typing import Sequence, Union
+
+import sqlalchemy as sa
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "7b60b08c6475"
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "tarefas_base",
+        sa.Column("nome", sa.String(length=255), nullable=False),
+        sa.Column("descricao", sa.String(length=255), nullable=False),
+        sa.Column("prazo_em_meses", sa.Integer(), nullable=False),
+        sa.Column(
+            "curso",
+            sa.Enum("MESTRADO", "DOUTORADO", name="cursoalunoenum"),
+            nullable=False,
+        ),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("created_at", sa.Date(), nullable=False),
+        sa.Column("updated_at", sa.Date(), nullable=False),
+        sa.Column("deleted_at", sa.Date(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_tarefas_base_curso"), "tarefas_base", ["curso"], unique=False
+    )
+    op.create_index(op.f("ix_tarefas_base_id"), "tarefas_base", ["id"], unique=False)
+    op.create_table(
+        "tipo_usuario",
+        sa.Column(
+            "titulo",
+            sa.Enum("COORDENADOR", "PROFESSOR", "ALUNO", name="tipousuarioenum"),
+            nullable=False,
+        ),
+        sa.Column("descricao", sa.String(length=255), nullable=False),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("created_at", sa.Date(), nullable=False),
+        sa.Column("updated_at", sa.Date(), nullable=False),
+        sa.Column("deleted_at", sa.Date(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(
+        op.f("ix_tipo_usuario_descricao"), "tipo_usuario", ["descricao"], unique=False
+    )
+    op.create_index(op.f("ix_tipo_usuario_id"), "tipo_usuario", ["id"], unique=False)
+    op.create_index(
+        op.f("ix_tipo_usuario_titulo"), "tipo_usuario", ["titulo"], unique=False
+    )
+    op.create_table(
+        "usuarios",
+        sa.Column("nome", sa.String(length=255), nullable=False),
+        sa.Column("email", sa.String(length=255), nullable=False),
+        sa.Column("senha_hash", sa.String(length=255), nullable=False),
+        sa.Column("new_password_token", sa.String(length=255), nullable=True),
+        sa.Column("tipo_usuario_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("created_at", sa.Date(), nullable=False),
+        sa.Column("updated_at", sa.Date(), nullable=False),
+        sa.Column("deleted_at", sa.Date(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["tipo_usuario_id"],
+            ["tipo_usuario.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_usuarios_email"), "usuarios", ["email"], unique=True)
+    op.create_index(op.f("ix_usuarios_id"), "usuarios", ["id"], unique=False)
+    op.create_index(op.f("ix_usuarios_nome"), "usuarios", ["nome"], unique=False)
+    op.create_index(
+        op.f("ix_usuarios_senha_hash"), "usuarios", ["senha_hash"], unique=False
+    )
+    op.create_index(
+        op.f("ix_usuarios_tipo_usuario_id"),
+        "usuarios",
+        ["tipo_usuario_id"],
+        unique=False,
+    )
+    op.create_table(
+        "professores",
+        sa.Column("usuario_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("created_at", sa.Date(), nullable=False),
+        sa.Column("updated_at", sa.Date(), nullable=False),
+        sa.Column("deleted_at", sa.Date(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["usuario_id"],
+            ["usuarios.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_professores_id"), "professores", ["id"], unique=False)
+    op.create_index(
+        op.f("ix_professores_usuario_id"), "professores", ["usuario_id"], unique=False
+    )
+    op.create_table(
+        "alunos",
+        sa.Column("cpf", sa.String(length=14), nullable=False),
+        sa.Column("telefone", sa.String(length=23), nullable=False),
+        sa.Column("matricula", sa.String(length=20), nullable=False),
+        sa.Column("lattes", sa.String(length=255), nullable=True),
+        sa.Column(
+            "curso",
+            sa.Enum("MESTRADO", "DOUTORADO", name="cursoalunoenum"),
+            nullable=False,
+        ),
+        sa.Column("data_ingresso", sa.Date(), nullable=False),
+        sa.Column("data_qualificacao", sa.Date(), nullable=True),
+        sa.Column("data_defesa", sa.Date(), nullable=True),
+        sa.Column("orientador_id", sa.Integer(), nullable=True),
+        sa.Column("usuario_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("created_at", sa.Date(), nullable=False),
+        sa.Column("updated_at", sa.Date(), nullable=False),
+        sa.Column("deleted_at", sa.Date(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["orientador_id"],
+            ["professores.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["usuario_id"],
+            ["usuarios.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_alunos_cpf"), "alunos", ["cpf"], unique=True)
+    op.create_index(op.f("ix_alunos_curso"), "alunos", ["curso"], unique=False)
+    op.create_index(
+        op.f("ix_alunos_data_defesa"), "alunos", ["data_defesa"], unique=False
+    )
+    op.create_index(
+        op.f("ix_alunos_data_ingresso"), "alunos", ["data_ingresso"], unique=False
+    )
+    op.create_index(
+        op.f("ix_alunos_data_qualificacao"),
+        "alunos",
+        ["data_qualificacao"],
+        unique=False,
+    )
+    op.create_index(op.f("ix_alunos_id"), "alunos", ["id"], unique=False)
+    op.create_index(op.f("ix_alunos_lattes"), "alunos", ["lattes"], unique=False)
+    op.create_index(op.f("ix_alunos_matricula"), "alunos", ["matricula"], unique=True)
+    op.create_index(
+        op.f("ix_alunos_orientador_id"), "alunos", ["orientador_id"], unique=False
+    )
+    op.create_index(op.f("ix_alunos_telefone"), "alunos", ["telefone"], unique=True)
+    op.create_table(
+        "solicitacoes",
+        sa.Column("aluno_id", sa.Integer(), nullable=False),
+        sa.Column("professor_id", sa.Integer(), nullable=False),
+        sa.Column(
+            "status",
+            sa.Enum("PENDENTE", "ACEITA", "RECUSADA", name="statussolicitacaoenum"),
+            nullable=False,
+        ),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("created_at", sa.Date(), nullable=False),
+        sa.Column("updated_at", sa.Date(), nullable=False),
+        sa.Column("deleted_at", sa.Date(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["aluno_id"],
+            ["alunos.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["professor_id"],
+            ["professores.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_solicitacoes_id"), "solicitacoes", ["id"], unique=False)
+    op.create_index(
+        op.f("ix_solicitacoes_professor_id"),
+        "solicitacoes",
+        ["professor_id"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_solicitacoes_status"), "solicitacoes", ["status"], unique=False
+    )
+    op.create_table(
+        "tarefas",
+        sa.Column("nome", sa.String(length=255), nullable=False),
+        sa.Column("descricao", sa.String(length=255), nullable=False),
+        sa.Column("data_prazo", sa.Date(), nullable=False),
+        sa.Column("data_ultima_notificacao", sa.Date(), nullable=True),
+        sa.Column("data_conclusao", sa.Date(), nullable=True),
+        sa.Column("aluno_id", sa.Integer(), nullable=False),
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("created_at", sa.Date(), nullable=False),
+        sa.Column("updated_at", sa.Date(), nullable=False),
+        sa.Column("deleted_at", sa.Date(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["aluno_id"],
+            ["alunos.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_tarefas_aluno_id"), "tarefas", ["aluno_id"], unique=False)
+    op.create_index(op.f("ix_tarefas_id"), "tarefas", ["id"], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f("ix_tarefas_id"), table_name="tarefas")
+    op.drop_index(op.f("ix_tarefas_aluno_id"), table_name="tarefas")
+    op.drop_table("tarefas")
+    op.drop_index(op.f("ix_solicitacoes_status"), table_name="solicitacoes")
+    op.drop_index(op.f("ix_solicitacoes_professor_id"), table_name="solicitacoes")
+    op.drop_index(op.f("ix_solicitacoes_id"), table_name="solicitacoes")
+    op.drop_table("solicitacoes")
+    op.drop_index(op.f("ix_alunos_telefone"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_orientador_id"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_matricula"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_lattes"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_id"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_data_qualificacao"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_data_ingresso"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_data_defesa"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_curso"), table_name="alunos")
+    op.drop_index(op.f("ix_alunos_cpf"), table_name="alunos")
+    op.drop_table("alunos")
+    op.drop_index(op.f("ix_professores_usuario_id"), table_name="professores")
+    op.drop_index(op.f("ix_professores_id"), table_name="professores")
+    op.drop_table("professores")
+    op.drop_index(op.f("ix_usuarios_tipo_usuario_id"), table_name="usuarios")
+    op.drop_index(op.f("ix_usuarios_senha_hash"), table_name="usuarios")
+    op.drop_index(op.f("ix_usuarios_nome"), table_name="usuarios")
+    op.drop_index(op.f("ix_usuarios_id"), table_name="usuarios")
+    op.drop_index(op.f("ix_usuarios_email"), table_name="usuarios")
+    op.drop_table("usuarios")
+    op.drop_index(op.f("ix_tipo_usuario_titulo"), table_name="tipo_usuario")
+    op.drop_index(op.f("ix_tipo_usuario_id"), table_name="tipo_usuario")
+    op.drop_index(op.f("ix_tipo_usuario_descricao"), table_name="tipo_usuario")
+    op.drop_table("tipo_usuario")
+    op.drop_index(op.f("ix_tarefas_base_id"), table_name="tarefas_base")
+    op.drop_index(op.f("ix_tarefas_base_curso"), table_name="tarefas_base")
+    op.drop_table("tarefas_base")
+    # ### end Alembic commands ###
