--- src/api/database/alembic/versions/c15640acc34d_add_tabela_usuarios_e_notificacoes.py
@@ -1,286 +1,0 @@
-"""Add tabela usuarios e notificacoes
-
-Revision ID: c15640acc34d
-Revises: 0c9f61bb86b0
-Create Date: 2024-06-17 01:15:17.302439
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = "c15640acc34d"
-down_revision: Union[str, None] = "0c9f61bb86b0"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "tipo_usuario",
-        sa.Column(
-            "tipo_usuario",
-            sa.Enum("COORDENADOR", "PROFESSOR", "ALUNO", name="tipousuarioenum"),
-            nullable=False,
-        ),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("created_at", sa.Date(), nullable=False),
-        sa.Column("updated_at", sa.Date(), nullable=False),
-        sa.Column("deleted_at", sa.Date(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_tipo_usuario_id"), "tipo_usuario", ["id"], unique=False)
-    op.create_index(
-        op.f("ix_tipo_usuario_tipo_usuario"),
-        "tipo_usuario",
-        ["tipo_usuario"],
-        unique=False,
-    )
-    op.create_table(
-        "usuarios",
-        sa.Column("nome", sa.String(length=255), nullable=False),
-        sa.Column("email", sa.String(length=255), nullable=False),
-        sa.Column("senha_hash", sa.String(length=255), nullable=False),
-        sa.Column("new_password_token", sa.String(length=255), nullable=True),
-        sa.Column("tipo_usuario_id", sa.Integer(), nullable=False),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("created_at", sa.Date(), nullable=False),
-        sa.Column("updated_at", sa.Date(), nullable=False),
-        sa.Column("deleted_at", sa.Date(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_usuarios_email"), "usuarios", ["email"], unique=True)
-    op.create_index(op.f("ix_usuarios_id"), "usuarios", ["id"], unique=False)
-    op.create_index(op.f("ix_usuarios_nome"), "usuarios", ["nome"], unique=False)
-    op.create_index(
-        op.f("ix_usuarios_senha_hash"), "usuarios", ["senha_hash"], unique=False
-    )
-    op.create_index(
-        op.f("ix_usuarios_tipo_usuario_id"),
-        "usuarios",
-        ["tipo_usuario_id"],
-        unique=False,
-    )
-    op.create_table(
-        "solicitacoes",
-        sa.Column("aluno_id", sa.Integer(), nullable=False),
-        sa.Column("professor_id", sa.Integer(), nullable=False),
-        sa.Column(
-            "status",
-            sa.Enum("PENDENTE", "ACEITA", "RECUSADA", name="statussolicitacaoenum"),
-            nullable=False,
-        ),
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("created_at", sa.Date(), nullable=False),
-        sa.Column("updated_at", sa.Date(), nullable=False),
-        sa.Column("deleted_at", sa.Date(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["aluno_id"],
-            ["alunos.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["professor_id"],
-            ["professores.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_solicitacoes_id"), "solicitacoes", ["id"], unique=False)
-    op.create_index(
-        op.f("ix_solicitacoes_status"), "solicitacoes", ["status"], unique=False
-    )
-    op.add_column("alunos", sa.Column("created_at", sa.Date(), nullable=False))
-    op.add_column("alunos", sa.Column("updated_at", sa.Date(), nullable=False))
-    op.add_column("alunos", sa.Column("deleted_at", sa.Date(), nullable=True))
-    op.alter_column(
-        "alunos",
-        "curso",
-        existing_type=mysql.ENUM("M", "D", collation="utf8mb4_unicode_ci"),
-        type_=sa.Enum("MESTRADO", "DOUTORADO", name="cursoalunoenum"),
-        existing_nullable=False,
-    )
-    op.drop_index("ix_alunos_email", table_name="alunos")
-    op.drop_index("ix_alunos_nome", table_name="alunos")
-    op.drop_index("ix_alunos_senha_hash", table_name="alunos")
-    op.drop_column("alunos", "new_password_token")
-    op.drop_column("alunos", "nome")
-    op.drop_column("alunos", "email")
-    op.drop_column("alunos", "senha_hash")
-    op.add_column("professores", sa.Column("usuario_id", sa.Integer(), nullable=False))
-    op.add_column("professores", sa.Column("created_at", sa.Date(), nullable=False))
-    op.add_column("professores", sa.Column("updated_at", sa.Date(), nullable=False))
-    op.add_column("professores", sa.Column("deleted_at", sa.Date(), nullable=True))
-    op.drop_index("ix_professores_email", table_name="professores")
-    op.drop_index("ix_professores_nome", table_name="professores")
-    op.create_foreign_key(None, "professores", "usuarios", ["usuario_id"], ["id"])
-    op.drop_column("professores", "nome")
-    op.drop_column("professores", "role")
-    op.drop_column("professores", "new_password_token")
-    op.drop_column("professores", "email")
-    op.drop_column("professores", "senha_hash")
-    op.add_column(
-        "tarefas", sa.Column("data_ultima_notificacao", sa.Date(), nullable=True)
-    )
-    op.add_column("tarefas", sa.Column("created_at", sa.Date(), nullable=False))
-    op.add_column("tarefas", sa.Column("updated_at", sa.Date(), nullable=False))
-    op.add_column("tarefas", sa.Column("deleted_at", sa.Date(), nullable=True))
-    op.create_index(op.f("ix_tarefas_id"), "tarefas", ["id"], unique=False)
-    op.drop_column("tarefas", "completada")
-    op.drop_column("tarefas", "last_notified")
-    op.add_column("tarefas_base", sa.Column("created_at", sa.Date(), nullable=False))
-    op.add_column("tarefas_base", sa.Column("updated_at", sa.Date(), nullable=False))
-    op.add_column("tarefas_base", sa.Column("deleted_at", sa.Date(), nullable=True))
-    op.alter_column(
-        "tarefas_base",
-        "descricao",
-        existing_type=mysql.TEXT(collation="utf8mb4_unicode_ci"),
-        type_=sa.String(length=255),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "tarefas_base",
-        "curso",
-        existing_type=mysql.ENUM("M", "D", collation="utf8mb4_unicode_ci"),
-        type_=sa.Enum("MESTRADO", "DOUTORADO", name="cursoalunoenum"),
-        existing_nullable=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "tarefas_base",
-        "curso",
-        existing_type=sa.Enum("MESTRADO", "DOUTORADO", name="cursoalunoenum"),
-        type_=mysql.ENUM("M", "D", collation="utf8mb4_unicode_ci"),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "tarefas_base",
-        "descricao",
-        existing_type=sa.String(length=255),
-        type_=mysql.TEXT(collation="utf8mb4_unicode_ci"),
-        existing_nullable=False,
-    )
-    op.drop_column("tarefas_base", "deleted_at")
-    op.drop_column("tarefas_base", "updated_at")
-    op.drop_column("tarefas_base", "created_at")
-    op.add_column("tarefas", sa.Column("last_notified", sa.DATE(), nullable=True))
-    op.add_column(
-        "tarefas",
-        sa.Column("completada", mysql.INTEGER(), autoincrement=False, nullable=False),
-    )
-    op.drop_index(op.f("ix_tarefas_id"), table_name="tarefas")
-    op.drop_column("tarefas", "deleted_at")
-    op.drop_column("tarefas", "updated_at")
-    op.drop_column("tarefas", "created_at")
-    op.drop_column("tarefas", "data_ultima_notificacao")
-    op.add_column(
-        "professores",
-        sa.Column(
-            "senha_hash",
-            mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=255),
-            nullable=False,
-        ),
-    )
-    op.add_column(
-        "professores",
-        sa.Column(
-            "email",
-            mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=255),
-            nullable=False,
-        ),
-    )
-    op.add_column(
-        "professores",
-        sa.Column(
-            "new_password_token",
-            mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=255),
-            nullable=True,
-        ),
-    )
-    op.add_column(
-        "professores",
-        sa.Column(
-            "role",
-            mysql.ENUM("orientador", "coordenador", collation="utf8mb4_unicode_ci"),
-            nullable=False,
-        ),
-    )
-    op.add_column(
-        "professores",
-        sa.Column(
-            "nome",
-            mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=255),
-            nullable=False,
-        ),
-    )
-    op.drop_constraint(None, "professores", type_="foreignkey")
-    op.create_index("ix_professores_nome", "professores", ["nome"], unique=False)
-    op.create_index("ix_professores_email", "professores", ["email"], unique=True)
-    op.drop_column("professores", "deleted_at")
-    op.drop_column("professores", "updated_at")
-    op.drop_column("professores", "created_at")
-    op.drop_column("professores", "usuario_id")
-    op.add_column(
-        "alunos",
-        sa.Column(
-            "senha_hash",
-            mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=255),
-            nullable=False,
-        ),
-    )
-    op.add_column(
-        "alunos",
-        sa.Column(
-            "email",
-            mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=255),
-            nullable=False,
-        ),
-    )
-    op.add_column(
-        "alunos",
-        sa.Column(
-            "nome",
-            mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=255),
-            nullable=False,
-        ),
-    )
-    op.add_column(
-        "alunos",
-        sa.Column(
-            "new_password_token",
-            mysql.VARCHAR(collation="utf8mb4_unicode_ci", length=255),
-            nullable=True,
-        ),
-    )
-    op.create_index("ix_alunos_senha_hash", "alunos", ["senha_hash"], unique=False)
-    op.create_index("ix_alunos_nome", "alunos", ["nome"], unique=False)
-    op.create_index("ix_alunos_email", "alunos", ["email"], unique=True)
-    op.alter_column(
-        "alunos",
-        "curso",
-        existing_type=sa.Enum("MESTRADO", "DOUTORADO", name="cursoalunoenum"),
-        type_=mysql.ENUM("M", "D", collation="utf8mb4_unicode_ci"),
-        existing_nullable=False,
-    )
-    op.drop_column("alunos", "deleted_at")
-    op.drop_column("alunos", "updated_at")
-    op.drop_column("alunos", "created_at")
-    op.drop_index(op.f("ix_solicitacoes_status"), table_name="solicitacoes")
-    op.drop_index(op.f("ix_solicitacoes_id"), table_name="solicitacoes")
-    op.drop_table("solicitacoes")
-    op.drop_index(op.f("ix_usuarios_tipo_usuario_id"), table_name="usuarios")
-    op.drop_index(op.f("ix_usuarios_senha_hash"), table_name="usuarios")
-    op.drop_index(op.f("ix_usuarios_nome"), table_name="usuarios")
-    op.drop_index(op.f("ix_usuarios_id"), table_name="usuarios")
-    op.drop_index(op.f("ix_usuarios_email"), table_name="usuarios")
-    op.drop_table("usuarios")
-    op.drop_index(op.f("ix_tipo_usuario_tipo_usuario"), table_name="tipo_usuario")
-    op.drop_index(op.f("ix_tipo_usuario_id"), table_name="tipo_usuario")
-    op.drop_table("tipo_usuario")
-    # ### end Alembic commands ###
