--- src/api/database/alembic/versions/c67dbd4a354c_renomear_coluna.py
@@ -1,56 +1,0 @@
-"""renomear coluna
-
-Revision ID: c67dbd4a354c
-Revises: 713e6a02abcb
-Create Date: 2024-06-17 03:50:06.743905
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-from alembic import op
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision: str = "c67dbd4a354c"
-down_revision: Union[str, None] = "713e6a02abcb"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "tipo_usuario",
-        sa.Column(
-            "titulo",
-            sa.Enum("COORDENADOR", "PROFESSOR", "ALUNO", name="tipousuarioenum"),
-            nullable=False,
-        ),
-    )
-    op.drop_index("ix_tipo_usuario_tipo_usuario", table_name="tipo_usuario")
-    op.create_index(
-        op.f("ix_tipo_usuario_titulo"), "tipo_usuario", ["titulo"], unique=False
-    )
-    op.drop_column("tipo_usuario", "tipo_usuario")
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "tipo_usuario",
-        sa.Column(
-            "tipo_usuario",
-            mysql.ENUM(
-                "COORDENADOR", "PROFESSOR", "ALUNO", collation="utf8mb4_unicode_ci"
-            ),
-            nullable=False,
-        ),
-    )
-    op.drop_index(op.f("ix_tipo_usuario_titulo"), table_name="tipo_usuario")
-    op.create_index(
-        "ix_tipo_usuario_tipo_usuario", "tipo_usuario", ["tipo_usuario"], unique=False
-    )
-    op.drop_column("tipo_usuario", "titulo")
-    # ### end Alembic commands ###
